Sample n rows from a table

sample_n() and sample_frac() have been superseded in favour of slice_sample(). While they will not be deprecated in the near future, retirement means that we will only perform critical bug fixes, so we recommend moving to the newer alternative.

These functions were superseded because we realised it was more convenient to have two mutually exclusive arguments to one function, rather than two separate functions. This also made it to clean up a few other smaller design issues with sample_n()/sample_frac:

The connection to slice() was not obvious.

The name of the first argument, tbl, is inconsistent with other single table verbs which use .data.

The size argument uses tidy evaluation, which is surprising and undocumented.

It was easier to remove the deprecated .env argument.

... was in a suboptimal position.

Usage

# S3 method for SpatialExperiment
sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for SpatialExperiment
sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

Arguments

tbl

A data.frame.

size

<tidy-select> For sample_n(), the number of rows to select. For sample_frac(), the fraction of rows to select. If tbl is grouped, size applies to each group.

replace

Sample with or without replacement?

weight

<tidy-select> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env

DEPRECATED.

...

ignored

Value

tidySpatialExperiment

Examples

example(read10xVisium)
#> 
#> rd10xV> dir <- system.file(
#> rd10xV+   file.path("extdata", "10xVisium"), 
#> rd10xV+   package = "SpatialExperiment")
#> 
#> rd10xV> sample_ids <- c("section1", "section2")
#> 
#> rd10xV> samples <- file.path(dir, sample_ids, "outs")
#> 
#> rd10xV> list.files(samples[1])
#> [1] "raw_feature_bc_matrix" "spatial"              
#> 
#> rd10xV> list.files(file.path(samples[1], "spatial"))
#> [1] "scalefactors_json.json"    "tissue_lowres_image.png"  
#> [3] "tissue_positions_list.csv"
#> 
#> rd10xV> file.path(samples[1], "raw_feature_bc_matrix")
#> [1] "/Users/hutchison.w/Library/R/x86_64/4.3/library/SpatialExperiment/extdata/10xVisium/section1/outs/raw_feature_bc_matrix"
#> 
#> rd10xV> (spe <- read10xVisium(samples, sample_ids, 
#> rd10xV+   type = "sparse", data = "raw", 
#> rd10xV+   images = "lowres", load = FALSE))
#> # A SpatialExperiment-tibble abstraction: 99 × 7
#> # Features=50 | Cells=99 | Assays=counts
#>    .cell              in_tissue array_row array_col sample_id pxl_col_in_fullres
#>    <chr>              <lgl>         <int>     <int> <chr>                  <int>
#>  1 AAACAACGAATAGTTC-1 FALSE             0        16 section1                2312
#>  2 AAACAAGTATCTCCCA-1 TRUE             50       102 section1                8230
#>  3 AAACAATCTACTAGCA-1 TRUE              3        43 section1                4170
#>  4 AAACACCAATAACTGC-1 TRUE             59        19 section1                2519
#>  5 AAACAGAGCGACTCCT-1 TRUE             14        94 section1                7679
#>  6 AAACAGCTTTCAGAAG-1 FALSE            43         9 section1                1831
#>  7 AAACAGGGTCTATATT-1 FALSE            47        13 section1                2106
#>  8 AAACAGTGTTCCTGGG-1 FALSE            73        43 section1                4170
#>  9 AAACATGGTGAGAGGA-1 FALSE            62         0 section1                1212
#> 10 AAACATTTCCCGGATT-1 FALSE            61        97 section1                7886
#> # ℹ 89 more rows
#> # ℹ 1 more variable: pxl_row_in_fullres <int>
#> 
#> rd10xV> # base directory 'outs/' from Space Ranger can also be omitted
#> rd10xV> samples2 <- file.path(dir, sample_ids)
#> 
#> rd10xV> (spe2 <- read10xVisium(samples2, sample_ids, 
#> rd10xV+   type = "sparse", data = "raw", 
#> rd10xV+   images = "lowres", load = FALSE))
#> # A SpatialExperiment-tibble abstraction: 99 × 7
#> # Features=50 | Cells=99 | Assays=counts
#>    .cell              in_tissue array_row array_col sample_id pxl_col_in_fullres
#>    <chr>              <lgl>         <int>     <int> <chr>                  <int>
#>  1 AAACAACGAATAGTTC-1 FALSE             0        16 section1                2312
#>  2 AAACAAGTATCTCCCA-1 TRUE             50       102 section1                8230
#>  3 AAACAATCTACTAGCA-1 TRUE              3        43 section1                4170
#>  4 AAACACCAATAACTGC-1 TRUE             59        19 section1                2519
#>  5 AAACAGAGCGACTCCT-1 TRUE             14        94 section1                7679
#>  6 AAACAGCTTTCAGAAG-1 FALSE            43         9 section1                1831
#>  7 AAACAGGGTCTATATT-1 FALSE            47        13 section1                2106
#>  8 AAACAGTGTTCCTGGG-1 FALSE            73        43 section1                4170
#>  9 AAACATGGTGAGAGGA-1 FALSE            62         0 section1                1212
#> 10 AAACATTTCCCGGATT-1 FALSE            61        97 section1                7886
#> # ℹ 89 more rows
#> # ℹ 1 more variable: pxl_row_in_fullres <int>
#> 
#> rd10xV> # tabulate number of spots mapped to tissue
#> rd10xV> cd <- colData(spe)
#> 
#> rd10xV> table(
#> rd10xV+   in_tissue = cd$in_tissue, 
#> rd10xV+   sample_id = cd$sample_id)
#>          sample_id
#> in_tissue section1 section2
#>     FALSE       28       27
#>     TRUE        22       22
#> 
#> rd10xV> # view available images
#> rd10xV> imgData(spe)
#> DataFrame with 2 rows and 4 columns
#>     sample_id    image_id   data scaleFactor
#>   <character> <character> <list>   <numeric>
#> 1    section1      lowres   ####   0.0510334
#> 2    section2      lowres   ####   0.0510334
spe |>
    sample_n(10)
#> # A SpatialExperiment-tibble abstraction: 10 × 7
#> # Features=50 | Cells=10 | Assays=counts
#>    .cell              in_tissue array_row array_col sample_id pxl_col_in_fullres
#>    <chr>              <lgl>         <int>     <int> <chr>                  <int>
#>  1 AAACGGGTTGGTATCC-1 FALSE             1        23 section1                2794
#>  2 AAAGACTGGGCGCTTT-1 FALSE            29        15 section1                2243
#>  3 AAACTGCTGGCTCCAA-1 TRUE             45        67 section1                5821
#>  4 AAACAGAGCGACTCCT-1 TRUE             14        94 section1                7679
#>  5 AAAGCTTGCCTACATA-1 FALSE            26       122 section1                9605
#>  6 AAACCGTTCGTCCAGG-1 TRUE             52        42 section1                4101
#>  7 AAACTCGTGATATAAG-1 FALSE            23       113 section1                8986
#>  8 AAACAATCTACTAGCA-1 TRUE              3        43 section1                4170
#>  9 AAACGGGCGTACGGGT-1 FALSE            65        91 section1                7473
#> 10 AAAGTCGACCCTCAGT-1 TRUE             37        15 section1                2244
#> # ℹ 1 more variable: pxl_row_in_fullres <int>
spe |>
    sample_frac(0.1)
#> tidySpatialExperiment says: When sampling with replacement a data frame is returned for independent data analysis.
#> # A tibble: 20 × 7
#>    .cell              in_tissue array_row array_col sample_id pxl_col_in_fullres
#>    <chr>              <lgl>         <int>     <int> <chr>                  <int>
#>  1 AAACATGGTGAGAGGA-1 FALSE            62         0 section1                1212
#>  2 AAACCTAAGCAGCCGG-1 FALSE            65        83 section1                6922
#>  3 AAACCTAAGCAGCCGG-1 FALSE            65        83 section1                6922
#>  4 AAACTCGTGATATAAG-1 FALSE            23       113 section1                8986
#>  5 AAAGCTTGCCTACATA-1 FALSE            26       122 section1                9605
#>  6 AAAGGGCAGCTTGAAT-1 TRUE             24        26 section1                3000
#>  7 AAAGGGCAGCTTGAAT-1 TRUE             24        26 section1                3000
#>  8 AAAGGTAAGCTGTACC-1 FALSE            10       106 section1                8504
#>  9 AAAGGTCAACGACATG-1 FALSE             0       112 section1                8917
#> 10 AAAGTAGCATTGCTCA-1 TRUE             51        27 section1                3069
#> 11 AAACATGGTGAGAGGA-1 FALSE            62         0 section2                1212
#> 12 AAACCTAAGCAGCCGG-1 FALSE            65        83 section2                6922
#> 13 AAACCTAAGCAGCCGG-1 FALSE            65        83 section2                6922
#> 14 AAACTCGTGATATAAG-1 FALSE            23       113 section2                8986
#> 15 AAAGCTTGCCTACATA-1 FALSE            26       122 section2                9605
#> 16 AAAGGGCAGCTTGAAT-1 TRUE             24        26 section2                3000
#> 17 AAAGGGCAGCTTGAAT-1 TRUE             24        26 section2                3000
#> 18 AAAGGTAAGCTGTACC-1 FALSE            10       106 section2                8504
#> 19 AAAGGTCAACGACATG-1 FALSE             0       112 section2                8917
#> 20 AAAGTAGCATTGCTCA-1 TRUE             51        27 section2                3069
#> # ℹ 1 more variable: pxl_row_in_fullres <int>